requirements.txt
pip~=24.0
pillow~=10.3.0
filelock~=3.17.0
setuptools~=65.5.0
jinja2~=3.1.5
PyYAML~=6.0.2